+++ src/js/widgets-controller.js
@@ -178,7 +178,7 @@ define( function( require, exports, module ) {
      * @param {jQuery}         $els   The jQuery collection of elements that the widget has been instantiated on.
      */
     _setOptionChangeListener = function( widget, $els ) {
-        if ( $els.length > 0 && $els.prop( 'nodeName' ).toLowerCase() === 'select' ) {
+        if ( $els.length > 0 && widget.list ) {
             $els.on( 'changeoption', function() {
                 // update (itemselect) picker on which event was triggered because the options changed
                 $( this )[ widget.name ]( 'update' );